tests: try to check editing plugin
